backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
pandas>=2.0.0
numpy==1.26.4
openai>=1.12.0
python-dotenv==1.0.0
python-multipart==0.0.6
pydantic-settings==2.1.0
PyJWT==2.9.0

